Add multiple policy rules to policy form
