+++ index.html
@@ -1,10 +1,43 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/favicon.ico" />
+    <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vultisig Plugin</title>
+    <title>Plugin Marketplace</title>
+    <meta name="description" content="Plugin Marketplace" />
+    <meta name="keywords" content="Plugin Marketplace" />
+    <meta name="author" content="Vultisig" />
+    <link rel="author" href="https://vultisig.com" />
+
+    <link rel="apple-touch-icon" href="/favicon.ico" />
+    <link rel="apple-touch-startup-image" href="/favicon.ico" />
+    <meta name="apple-mobile-web-app-capable" content="yes" />
+    <meta name="apple-mobile-web-app-title" content="Plugin Marketplace" />
+    <meta
+      name="apple-mobile-web-app-status-bar-style"
+      content="black-translucent"
+    />
+
+    <meta property="og:title" content="Plugin Marketplace" />
+    <meta property="og:description" content="Plugin Marketplace" />
+    <meta property="og:url" content="https://store.vultisigplugin.app" />
+    <meta property="og:site_name" content="Plugin Marketplace" />
+    <meta property="og:type" content="website" />
+    <meta
+      property="og:image"
+      content="https://store.vultisigplugin.app/images/banner.jpg"
+    />
+
+    <meta name="twitter:title" content="Plugin Marketplace" />
+    <meta name="twitter:description" content="Plugin Marketplace" />
+    <meta name="twitter:site" content="https://store.vultisigplugin.app" />
+    <meta name="twitter:creator" content="Vultisig" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta
+      name="twitter:image"
+      content="https://store.vultisigplugin.app/images/banner.jpg"
+    />
   </head>
   <body>
     <div id="root"></div>
