+++ src/pages/not_found/index.tsx
@@ -2,7 +2,7 @@ import { Layout, Result } from "antd";
 import { useTheme } from "styled-components";
 
 import { Button } from "@/components/Button";
-import { Stack } from "@/components/Stack";
+import { HStack, VStack } from "@/components/Stack";
 import { useGoBack } from "@/hooks/useGoBack";
 import { routeTree } from "@/utils/constants/routes";
 
@@ -11,25 +11,32 @@ export const NotFoundPage = () => {
   const colors = useTheme();
 
   return (
-    <Stack
-      as={Layout}
-      $style={{
-        alignItems: "center",
-        backgroundColor: colors.bgPrimary.toHex(),
-        justifyContent: "center",
-        height: "100%",
-      }}
-    >
-      <Result
-        status="404"
-        title="404"
-        subTitle="Sorry, the page you visited does not exist."
-        extra={
-          <Button kind="primary" onClick={() => goBack(routeTree.root.path)}>
-            Back Home
-          </Button>
-        }
-      />
-    </Stack>
+    <VStack $style={{ alignItems: "center", flexGrow: "1" }}>
+      <VStack
+        as={Layout}
+        $style={{
+          alignItems: "center",
+          backgroundColor: colors.bgPrimary.toHex(),
+          justifyContent: "center",
+          height: "100%",
+        }}
+      >
+        <Result
+          status="404"
+          title="404"
+          subTitle="Sorry, the page you visited does not exist."
+          extra={
+            <HStack $style={{ justifyContent: "center" }}>
+              <Button
+                kind="primary"
+                onClick={() => goBack(routeTree.root.path)}
+              >
+                Back Home
+              </Button>
+            </HStack>
+          }
+        />
+      </VStack>
+    </VStack>
   );
 };
